Version 1.0.0-alpha
The alpha release of version 1.0.0 now contains documentation of the core API. The documentation is built with sphinx and formatted to work with readthedocs.

It also contains modifications to the API to make some things more concise and fix a few bugs.

Version 1.0.0 should be released once there are more examples in place.